Tests with emscripten

+++ src/libpumpkin/puff.h
@@ -27,6 +27,6 @@
  */
 
 int puff(unsigned char *dest,           /* pointer to destination pointer */
-         unsigned long *destlen,        /* amount of output space */
+         uint32_t *destlen,        /* amount of output space */
          const unsigned char *source,   /* pointer to source data pointer */
-         unsigned long *sourcelen);     /* amount of input available */
+         uint32_t *sourcelen);     /* amount of input available */
